pyproject.toml
[project]
name = "hf-mockapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-nested-admin>=4.1.6",
    "django-restapi>=0.1.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pyyaml>=6.0.0",
]
